detectors/requirements.txt
-i https://pypi.org/simple
click==8.1.7; python_version >= '3.7'
flask==3.0.2; python_version >= '3.8'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.3; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
sqlalchemy==2.0.25; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
werkzeug==3.0.1; python_version >= '3.8'
PyYAML==6.0.1; python_version >= '3.8'
